requirements.txt
streamlit
pytube
moviepy
numpy
openai-whisper
requests
